+++ src/main/java/eu/endermite/censura/filter/FilterCache.java
@@ -1,0 +1,6 @@
+package eu.endermite.censura.filter;
+
+public class FilterCache {
+	public String spammySpacesRemoved;
+	public String spammySpacesRemovedNoRepeat;
+}
